+++ libsrc/libQFieldValidator/Exceptions.cpp
