+++ libsrc/libQFieldValidator/Exceptions.h
