+++ libsrc/libQFieldValidator/Private/GenericValidators.cpp
@@ -27,7 +27,7 @@
 #include <QUrl>
 #include <QDateTime>
 
-#include "libsrc/PhoneNumberUtil.hpp"
+#include "PhoneNumberUtil.hpp"
 using namespace Targoman;
 
 namespace Validators {
