+++ libsrc/libQFieldValidator/QFieldValidator.cpp
